requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pymongo==4.6.1
python-multipart==0.0.6
httpx==0.26.0
python-dotenv==1.0.0
google-generativeai==0.7.2
pydub==0.25.1